refactor(py): refactor to align with hftbacktest refactoring.

+++ py-hftbacktest/src/depth.rs
@@ -6,37 +6,37 @@ use hftbacktest::{
 };
 
 #[no_mangle]
-pub extern "C" fn hashmapdepth_best_bid_tick(ptr: *const HashMapMarketDepth) -> i32 {
+pub extern "C" fn hashmapdepth_best_bid_tick(ptr: *const HashMapMarketDepth) -> i64 {
     let depth = unsafe { &*ptr };
     depth.best_bid_tick()
 }
 
 #[no_mangle]
-pub extern "C" fn hashmapdepth_best_ask_tick(ptr: *const HashMapMarketDepth) -> i32 {
+pub extern "C" fn hashmapdepth_best_ask_tick(ptr: *const HashMapMarketDepth) -> i64 {
     let depth = unsafe { &*ptr };
     depth.best_ask_tick()
 }
 
 #[no_mangle]
-pub extern "C" fn hashmapdepth_best_bid(ptr: *const HashMapMarketDepth) -> f32 {
+pub extern "C" fn hashmapdepth_best_bid(ptr: *const HashMapMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.best_bid()
 }
 
 #[no_mangle]
-pub extern "C" fn hashmapdepth_best_ask(ptr: *const HashMapMarketDepth) -> f32 {
+pub extern "C" fn hashmapdepth_best_ask(ptr: *const HashMapMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.best_ask()
 }
 
 #[no_mangle]
-pub extern "C" fn hashmapdepth_tick_size(ptr: *const HashMapMarketDepth) -> f32 {
+pub extern "C" fn hashmapdepth_tick_size(ptr: *const HashMapMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.tick_size()
 }
 
 #[no_mangle]
-pub extern "C" fn hashmapdepth_lot_size(ptr: *const HashMapMarketDepth) -> f32 {
+pub extern "C" fn hashmapdepth_lot_size(ptr: *const HashMapMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.lot_size()
 }
@@ -44,8 +44,8 @@ pub extern "C" fn hashmapdepth_lot_size(ptr: *const HashMapMarketDepth) -> f32 {
 #[no_mangle]
 pub extern "C" fn hashmapdepth_bid_qty_at_tick(
     ptr: *const HashMapMarketDepth,
-    price_tick: i32,
-) -> f32 {
+    price_tick: i64,
+) -> f64 {
     let depth = unsafe { &*ptr };
     depth.bid_qty_at_tick(price_tick)
 }
@@ -53,8 +53,8 @@ pub extern "C" fn hashmapdepth_bid_qty_at_tick(
 #[no_mangle]
 pub extern "C" fn hashmapdepth_ask_qty_at_tick(
     ptr: *const HashMapMarketDepth,
-    price_tick: i32,
-) -> f32 {
+    price_tick: i64,
+) -> f64 {
     let depth = unsafe { &*ptr };
     depth.ask_qty_at_tick(price_tick)
 }
@@ -81,37 +81,37 @@ pub extern "C" fn hashmapdepth_snapshot_free(event_ptr: *mut Event, len: usize) 
 }
 
 #[no_mangle]
-pub extern "C" fn roivecdepth_best_bid_tick(ptr: *const ROIVectorMarketDepth) -> i32 {
+pub extern "C" fn roivecdepth_best_bid_tick(ptr: *const ROIVectorMarketDepth) -> i64 {
     let depth = unsafe { &*ptr };
     depth.best_bid_tick()
 }
 
 #[no_mangle]
-pub extern "C" fn roivecdepth_best_ask_tick(ptr: *const ROIVectorMarketDepth) -> i32 {
+pub extern "C" fn roivecdepth_best_ask_tick(ptr: *const ROIVectorMarketDepth) -> i64 {
     let depth = unsafe { &*ptr };
     depth.best_ask_tick()
 }
 
 #[no_mangle]
-pub extern "C" fn roivecdepth_best_bid(ptr: *const ROIVectorMarketDepth) -> f32 {
+pub extern "C" fn roivecdepth_best_bid(ptr: *const ROIVectorMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.best_bid()
 }
 
 #[no_mangle]
-pub extern "C" fn roivecdepth_best_ask(ptr: *const ROIVectorMarketDepth) -> f32 {
+pub extern "C" fn roivecdepth_best_ask(ptr: *const ROIVectorMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.best_ask()
 }
 
 #[no_mangle]
-pub extern "C" fn roivecdepth_tick_size(ptr: *const ROIVectorMarketDepth) -> f32 {
+pub extern "C" fn roivecdepth_tick_size(ptr: *const ROIVectorMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.tick_size()
 }
 
 #[no_mangle]
-pub extern "C" fn roivecdepth_lot_size(ptr: *const ROIVectorMarketDepth) -> f32 {
+pub extern "C" fn roivecdepth_lot_size(ptr: *const ROIVectorMarketDepth) -> f64 {
     let depth = unsafe { &*ptr };
     depth.lot_size()
 }
@@ -119,8 +119,8 @@ pub extern "C" fn roivecdepth_lot_size(ptr: *const ROIVectorMarketDepth) -> f32 
 #[no_mangle]
 pub extern "C" fn roivecdepth_bid_qty_at_tick(
     ptr: *const ROIVectorMarketDepth,
-    price_tick: i32,
-) -> f32 {
+    price_tick: i64,
+) -> f64 {
     let depth = unsafe { &*ptr };
     depth.bid_qty_at_tick(price_tick)
 }
@@ -128,8 +128,8 @@ pub extern "C" fn roivecdepth_bid_qty_at_tick(
 #[no_mangle]
 pub extern "C" fn roivecdepth_ask_qty_at_tick(
     ptr: *const ROIVectorMarketDepth,
-    price_tick: i32,
-) -> f32 {
+    price_tick: i64,
+) -> f64 {
     let depth = unsafe { &*ptr };
     depth.ask_qty_at_tick(price_tick)
 }
@@ -138,7 +138,7 @@ pub extern "C" fn roivecdepth_ask_qty_at_tick(
 pub extern "C" fn roivecdepth_bid_depth(
     ptr: *const ROIVectorMarketDepth,
     len: *mut usize,
-) -> *const f32 {
+) -> *const f64 {
     let depth = unsafe { &*ptr };
     unsafe { *len = depth.bid_depth().len() }
     depth.bid_depth().as_ptr()
@@ -148,7 +148,7 @@ pub extern "C" fn roivecdepth_bid_depth(
 pub extern "C" fn roivecdepth_ask_depth(
     ptr: *const ROIVectorMarketDepth,
     len: *mut usize,
-) -> *const f32 {
+) -> *const f64 {
     let depth = unsafe { &*ptr };
     unsafe { *len = depth.ask_depth().len() }
     depth.ask_depth().as_ptr()
